+++ cpp/src/communicator/ucxx_utils.cpp
@@ -53,7 +53,7 @@ std::shared_ptr<UCXX> init_using_mpi(MPI_Comm mpi_comm) {
     std::string root_worker_address_str{};
     std::shared_ptr<UCXX> comm;
     if (rank == 0) {
-        auto ucxx_initialized_rank = init(nullptr, nranks);
+        auto ucxx_initialized_rank = init(nullptr, static_cast<std::uint32_t>(nranks));
         comm = std::make_shared<UCXX>(std::move(ucxx_initialized_rank));
 
         root_listener_address = comm->listener_address();
@@ -66,7 +66,8 @@ std::shared_ptr<UCXX> init_using_mpi(MPI_Comm mpi_comm) {
     if (rank != 0) {
         auto root_worker_address =
             ::ucxx::createAddressFromString(root_worker_address_str);
-        auto ucxx_initialized_rank = init(nullptr, nranks, root_worker_address);
+        auto ucxx_initialized_rank =
+            init(nullptr, static_cast<std::uint32_t>(nranks), root_worker_address);
         comm = std::make_shared<UCXX>(std::move(ucxx_initialized_rank));
     }
